requirements.txt
azure-functions>=1.18.0
requests>=2.31.0
python-dateutil>=2.8.2
